requirements.txt
anthropic
langchain
langchain-community
pypdf
faiss-cpu
sentence-transformers
streamlit
python-dotenv
